move tests to a dev menu
