+++ baby-gru/src/components/MoorhenNavBar.js
@@ -7,6 +7,7 @@ import { MoorhenViewMenu } from './MoorhenViewMenu';
 import { MoorhenLigandMenu } from './MoorhenLigandMenu';
 import { MoorhenEditMenu } from './MoorhenEditMenu';
 import { MoorhenHelpMenu } from './MoorhenHelpMenu';
+import { MoorhenDevMenu } from './MoorhenDevMenu';
 import { SaveOutlined } from '@mui/icons-material';
 
 export const MoorhenNavBar = forwardRef((props, ref) => {
@@ -33,6 +34,7 @@ export const MoorhenNavBar = forwardRef((props, ref) => {
                         <MoorhenHistoryMenu dropdownId="History" {...collectedProps} />
                         <MoorhenPreferencesMenu dropdownId="Preferences" {...collectedProps} />
                         <MoorhenHelpMenu dropdownId="Help" {...collectedProps}/>
+                        {props.devMode && <MoorhenDevMenu dropdownId="Dev" {...collectedProps}/>}
                         {props.extraNavBarMenus && props.extraNavBarMenus.map(menu => menu)}
                     </Nav>
                 </Navbar.Collapse>
